ApiException: make Uri property's setter internal
This nitpick highlights the fact that ApiException is sadly not completely
immutable. However if we mark it internal it's a bit safer (as no consumers
of SD.Api.Wrapper library will be able to do evil impure things with it..).

Not sending this as a pull-request because:
- It's a 1-line change (easy revert if you don't agree).
- AnthonySteele has proposed a PR that moves all files in the Utility
namespace, so I'm just afraid that my PR would have conflicts and it would
slow down the process a lot for such a tiny change.KTHXBYE! :)