FluentApi: introduce 'Seal' proof of concept.
This Seal() new method would create a new Request object which would be
the only one allowed to call Please(). This way the parameters it handles
for the request cannot be modified if any thread comes along expecting
to reuse the same object to do a similar call but with different parameters,
like could happen before.