Refactored to remove UriSigner. In order to do this I had to invert a switch statement that selected which RequestHandler to use based on your request method, so now the RequestCoordinator ctor signature has changed to take an IEnumerable<RequestHandler>. This may cause a breaking change to the way you instantiate an instance of FLuentApi, but if your using a decent IoC framework you shouldnt have a problem.
Ive included a static RequestHandlerFactory that sets up the concrete Get and Post request handlers for now.

THis work is necesary not only for making replaceing the OAuth with the Crenna library, but also in order to support requestng the api with non GET and POST verbs in the future.